+++ src/gui/mod.rs
@@ -4,7 +4,7 @@ mod update;
 
 use crate::VERSION;
 use ajour_core::{
-    addon::Addon,
+    addon::{Addon, ReleaseChannel},
     config::{load_config, Config, Flavor},
     error::ClientError,
     fs::PersistentData,
@@ -69,6 +69,7 @@ pub enum Message {
     ParsedAddons((Flavor, Result<Vec<Addon>>)),
     UpdateFingerprint((String, Result<()>)),
     ThemeSelected(String),
+    ReleaseChannelSelected(ReleaseChannel),
     ThemesLoaded(Vec<Theme>),
     UnpackedAddon((String, Result<()>)),
     UpdateWowDirectory(Option<PathBuf>),